Add masks for POWER cpu features.
This commit add masks for get POWER cpu features. Those masks was taken from
cputable and will be used to parse the result from getauxval. Also this change
getauxval related functions to return uint32_t since the top 64 seems unused.

Notice that maybe we don't need all masks or we maybe missing some of them.

+++ src/powerpc/linux/hwcap.c
@@ -19,17 +19,17 @@
 #include <sys/auxv.h>
 
 #if CPUINFO_MOCK
-	static uint64_t mock_hwcap = 0;
-	void cpuinfo_set_hwcap(uint64_t hwcap) {
+	static uint32_t mock_hwcap = 0;
+	void cpuinfo_set_hwcap(uint32_t hwcap) {
 		mock_hwcap = hwcap;
 	}
 #endif
 
-uint64_t cpuinfo_powerpc_linux_hwcap_from_getauxval(void)
+uint32_t cpuinfo_powerpc_linux_hwcap_from_getauxval(void)
 {
 	#if CPUINFO_MOCK
 		return mock_hwcap;
 	#else
-		return getauxval(AT_HWCAP);
+		return (uint32_t) getauxval(AT_HWCAP);
 	#endif
 }
